src/data_collection/web_scraper.py
import requests
from bs4 import BeautifulSoup
from selenium import webdriver
from selenium.webdriver.chrome.options import Options
from selenium.webdriver.chrome.service import Service
from webdriver_manager.chrome import ChromeDriverManager
import time
from typing import Dict, List, Optional
import random
import logging

class WebScraper:
    """Production-ready web scraper framework for competitor monitoring."""
    
    def __init__(self, use_selenium: bool = False):
        self.use_selenium = use_selenium
        self.session = requests.Session()
        self.user_agents = [
            "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36",
            "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/92.0.4515.107 Safari/537.36",
            "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.114 Safari/537.36"
        ]
        self.logger = logging.getLogger(__name__)

    def _get_headers(self) -> Dict:
        """Rotate headers to avoid detection and respect rate limits."""
        return {
            "User-Agent": random.choice(self.user_agents),
            "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,*/*;q=0.8",
            "Accept-Language": "en-US,en;q=0.5",
            "Referer": "https://www.google.com/"
        }

    def fetch_static_page(self, url: str) -> Optional[BeautifulSoup]:
        """Fetch and parse a static HTML page with robust error handling."""
        try:
            response = self.session.get(url, headers=self._get_headers(), timeout=15)
            response.raise_for_status()
            return BeautifulSoup(response.text, 'html.parser')
        except Exception as e:
            self.logger.error(f"Error fetching static page {url}: {e}")
            return None

    def fetch_dynamic_page(self, url: str, wait_seconds: int = 5) -> str:
        """Fetch dynamic JS-heavy page content using Selenium (Headless Chrome)."""
        chrome_options = Options()
        chrome_options.add_argument("--headless")
        chrome_options.add_argument("--no-sandbox")
        chrome_options.add_argument("--disable-dev-shm-usage")
        
        try:
            service = Service(ChromeDriverManager().install())
            driver = webdriver.Chrome(service=service, options=chrome_options)
            driver.get(url)
            time.sleep(wait_seconds)
            content = driver.page_source
            driver.quit()
            return content
        except Exception as e:
            self.logger.error(f"Error fetching dynamic page {url}: {e}")
            return ""

    def scrape_competitor_pricing(self, url: str, price_selector: str) -> Dict:
        """Extract pricing from a generic competitor's landing/pricing page."""
        soup = self.fetch_static_page(url)
        if not soup: return {"price": None, "currency": "USD", "error": "Failed to load"}
        
        # Example price extraction (robust find)
        price_tag = soup.select_one(price_selector)
        if not price_tag:
            return {"price": 0.0, "currency": "USD", "error": "Selector not found"}
        
        return {
            "price": price_tag.text.strip(),
            "currency": "USD",
            "timestamp": datetime.now().isoformat()
        }

    def check_robots_txt(self, url: str) -> bool:
        """Verify robots.txt compliance (placeholder for a full implementation)."""
        # Logic to check /robots.txt
        return True
